requirements.txt
git+https://github.com/Rapptz/discord.py@async
requests==2.24.0
discord.py==1.5.1